refactor: ImageAreaSection로 분리
