+++ src/utils/result.ts
@@ -17,3 +17,12 @@ export const calcProgress = (symbolStack: number) => {
   const currentLevel = getLevel(symbolStack);
   return ((symbolStack - currentLevel.min) / (currentLevel.max - currentLevel.min)) * 100;
 };
+
+const MIN_PERCENT = 3;
+
+export const getPercent = (props: { symbolStack: number; max: number; min: number }): number => {
+  const { symbolStack, max, min } = props;
+  const percent = (100 / (max - min)) * (symbolStack - min);
+
+  return Math.max(percent, MIN_PERCENT);
+};
